inference/requirements.txt
# Inference Dependencies
torch>=2.0.0
transformers>=4.35.0
numpy>=1.24.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
